Keep feature discovery and resolution logic separate
